third_party/blink/web_tests/platform/generic/http/tests/security/media-element-audio-source-node-redirect-expected.txt

CONSOLE WARNING: The ScriptProcessorNode is deprecated. Use AudioWorkletNode instead. (https://bit.ly/audio-worklet)
CONSOLE MESSAGE: MediaElementAudioSource outputs zeroes due to CORS access restrictions for http://127.0.0.1:8000/serviceworker/resources/redirect.php?Redirect=http://localhost:8000/security/resources/webaudio/laughter.wav
MediaElementAudioSourceNode with redirection

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS All samples correctly zeroed.
PASS successfullyParsed is true

TEST COMPLETE
